Overwrite problem
By avoiding allocation of an empty array (in stead, allocate a dummy array) and assigning values to an empty array an overwrite problem was solved.
In silup computed gotos were removed. They invoked warnings by tge compiler.
In makefile a source was moved to the list of modules (NB the current makefile does not use this distinction)